src/app/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
python-multipart
jinja2
python-docx
PyPDF2
aiofiles
